move utils.geometry package to instruments.
