+++ digicampipe/instrument/geometry.py
@@ -2,7 +2,7 @@ from cts_core import camera as cam
 import numpy as np
 import astropy.units as u
 from ctapipe.instrument.camera import CameraGeometry
-from ..instrument.camera import _find_neighbor_pixels
+from digicampipe.instrument.camera import _find_neighbor_pixels
 
 
 def find_pixel_positions(
