+++ save_container_test.py
@@ -1,11 +1,12 @@
 import astropy.units as u
 from cts_core.camera import Camera
+from digicampipe.calib.camera import random_triggers
 
 from digicampipe.calib import filter
-from digicampipe.calib.camera import random_triggers
+from digicampipe.instrument import geometry
 from digicampipe.io.containers import save_to_pickle_gz
 from digicampipe.io.event_stream import event_stream
-from digicampipe.utils import utils, geometry
+from digicampipe.utils import utils
 
 if __name__ == '__main__':
     # Data configuration
